requirements.txt
asgiref==3.4.1
click==8.0.1
colorama==0.4.4

h11==0.12.0
loguru==0.5.3

opencv-python==4.5.3.56
Pillow==8.3.2
pydantic==1.8.2

six==1.16.0

torch==1.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2

win32-setctime==1.0.3
thefuzz
python-Levenshtein
fastapi
starlette
gunicorn
uvicorn
numpy
python-multipart
omegaconf
einops
gradio
scipy
ocrd-fork-pylsd
